Updates to README to reflect BF GH App
BreakFree organization has a new GitHub app, `breakfree-github-agent-
app`, and a new repo, `break-free/github-agent-demo`. This allows for
continuous reuse of both, rather than relying on a new user to create
their own. The README has been updated to reflect these changes.

A `setup/requirements.txt` file was added to cover the required
packages for the Python environment. Eventually, a build script for a
Fedora-based Toolbox will be added to make the environment setup as
easy as possible and containerized.

Minor changes were also made to `custom_tools.py` and `main.py`.